correct_user correct_product_statusをbefore_actionに追加

+++ app/controllers/api/v1/trade_messages_controller.rb
@@ -1,6 +1,8 @@
 class Api::V1::TradeMessagesController < ApplicationController
     before_action :authenticate_api_user!
     before_action :set_product
+    before_action :correct_user
+    before_action :correct_product_status
 
     def create
         @message = current_api_user.trade_messages.build(trade_messages_params)
@@ -14,8 +16,10 @@ class Api::V1::TradeMessagesController < ApplicationController
 
     def destroy
         @message = TradeMessage.find_by(id: params[:id])
-        if @message.destroy
-            render json: @message
+        if  current_api_user == @message.user
+            if @message.destroy
+                render json: @message
+            end
         else
             head :unprocessable_entity
         end
@@ -30,5 +34,15 @@ class Api::V1::TradeMessagesController < ApplicationController
         def set_product
             @product = Product.find_by(id: params[:product_id])
             head :not_found unless @product
+        end
+
+        def correct_user
+            unless current_api_user == @product.seller || current_api_user == @product.buyer
+                head :forbidden
+            end
+        end
+
+        def correct_product_status
+            head :forbidden unless @product.status == "trade"
         end
 end
